\name{icomps-package}
\alias{icomps-package}
\alias{icomps}
\docType{package}
\title{
Calculates the ICOMP criterion for a linear model.
}
\description{
  This package calculates the ICOMP criterion value for linear models as well as the ICOMP variations, ICOMPr and ICOMPcv.
}
\details{
\tabular{ll}{
Package: \tab icomp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-07-03\cr
% License: \tab What license is it under?\cr
}

The function ICOMP calls ICOMP.lm, ICOMPr.lm and ICOMPcv.lm to calculate each criterions value and then presents the output in a data frame.
}
\author{
Jake Ferguson

Maintainer: Jake Ferguson <troutinthemilk@gmail.com>

}
\references{
Bozdogan, H. Haughton, D.M.A (1998). Information complexity criteria for regression models. Computation Statistics & Data Analysis 28: 51-76

 Ferguson et al (unpublished) ICOMP: An Information Criterion new to Ecology
}
\keyword{ package }
\keyword{ models }
\keyword{ regression }
%\seealso{}

