\name{ICOMP.lm}
\alias{ICOMP.lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ ICOMP }
\description{
Calculates the ICOMP criterion value for a linear model.
}
\usage{
ICOMP.lm(lm.model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lm.model}{ a linear model of class 'lm' }
}

\value{
Returns the ICOMP value
}
\references{ Bozdogan, H. Haughton, D.M.A (1998). Information complexity criteria for regression models. Computation Statistics & Data Analysis 28: 51-76
}
\author{Jake Ferguson}

%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\keyword{ models }
\keyword{ regression }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
