\name{ICOMPcv.lm}
\alias{ICOMPcv.lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ ICOMPcv }
\description{
Calculates the ICOMPcv criterion for a linear model. ICOMPcv is similar to ICOMP but operates on the coefficient of variation matrix.
}
\usage{
ICOMPcv.lm(lm.model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lm.model}{ a linear model of class 'lm' }
}
\details{

}
\value{
  Returns the ICOMPcv value

}
\references{ Ferguson et al (unpublished) ICOMP: An Information Criterion new to Ecology }
\author{ Jake Ferguson}

% \seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\keyword{ models }
\keyword{ regression }


