\name{ICOMPr.lm}
\alias{ICOMPr.lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ ICOMPr }
\description{
Calculates the ICOMPr criterion value. ICOMPr is similar to ICOMP but operates on the correlation matrix.
}
\usage{
ICOMPr.lm(lm.model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lm.model}{ a linear model of class 'lm' }
}
\value{
returns the ICOMPr criterion value.
}
\references{ Bozdogan, H. Haughton, D.M.A (1998). Information complexity criteria for regression models. Computation Statistics & Data Analysis 28: 51-76 }
\author{ Jake Ferguson }

%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\keyword{ models }
\keyword{ regression }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
