% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICOMP.lm.r, R/ICOMP.r
\name{ICOMP.lm}
\alias{ICOMP}
\alias{ICOMP.lm}
\title{ICOMP}
\usage{
\method{ICOMP}{lm}(object, ..., complexity = "C1")

ICOMP(object, ...)
}
\arguments{
\item{object}{A model for which there is a method for \code{stats::logLik()}.
Currently supports \code{lm} objects.}

\item{...}{Additional arguments.}

\item{complexity}{Complexity measure, as a string.  Valid options are "C1".}
}
\value{
The ICOMP value for the given model.
}
\description{
Generic function for calculating Bozdogan's information complexity
measure (ICOMP).  ICOMP is a likelihood-based criterion similar
to Akaike's Information Criterion (AIC).  As with AIC, 
a smaller ICOMP value corresponds to a better model.
}
\examples{
\dontrun{
### Modified from the example in ?AIC
lm1 <- lm(Fertility ~ . , data = swiss)
lm2 <- update(lm1, . ~ . -Examination)

icomp::ICOMP(lm1, lm2)
}

}
\author{
Drew Schmidt and Jake Ferguson
}
\references{
Bozdogan, H. Haughton, D.M.A (1998). Information complexity
criteria for regression models. Computation Statistics & Data Analysis 28:
51-76
}
\keyword{models}
\keyword{regression}

